RSS feed URL submission functionality added.

+++ Valley.RssReader.Core/Controllers/RssFeedSurfaceController.cs
@@ -1,9 +1,12 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Web.Mvc;
 using Umbraco.Core.Models;
 using Umbraco.Web.Mvc;
 using Valley.RssReader.Common.Models;
 using Valley.RssReader.Common.Services.Interfaces;
+using Valley.RssReader.Core.Exceptions;
 using Valley.RssReader.Core.Services.Interfaces;
 
 namespace Valley.RssReader.Core.Controllers
@@ -19,23 +22,53 @@ namespace Valley.RssReader.Core.Controllers
             _rssItemMappingService = rssItemMappingService;
         }
 
-        // GET
-        public ActionResult Index()
+        public ActionResult RssUrlForm()
         {
-            Services.ContentService.DeleteContentOfType(1060);
+            return PartialView("RssUrlForm", new RssFeedUrlViewModel());
+        }
 
-            foreach (RssItemViewModel rssItem in _rssItemMappingService.Map(_rssReaderService.Read(new Uri("http://www.nu.nl/rss/Algemeen"))))
+        public ActionResult RssFeed()
+        {
+            return PartialView("RssFeed");
+        }
+
+        [HttpPost]
+        public ActionResult ImportUrl(RssFeedUrlViewModel rssFeedUrl)
+        {
+            if (!ModelState.IsValid)
             {
-                IContent rssItemContent = Services.ContentService.CreateContent($"RSS Item {rssItem.Id}", 1061, "RssItem");
-                rssItemContent.SetValue("title", rssItem.Title);
-                rssItemContent.SetValue("description", rssItem.Description);
-                rssItemContent.SetValue("categories", rssItem.Categories);
-                rssItemContent.SetValue("date", rssItem.Date);
-                rssItemContent.SetValue("link", rssItem.Link);
-                Services.ContentService.SaveAndPublishWithStatus(rssItemContent);
+                TempData.Add("Failure", "Please enter a valid URL.");
+                return CurrentUmbracoPage();
             }
 
-            return PartialView("RssFeed");
+            try
+            {
+                IEnumerable<IContent> rssItems = _rssItemMappingService.Map(_rssReaderService.Read(new Uri(rssFeedUrl.Url))).Select(m =>
+                {
+                    IContent rssItem = Services.ContentService.CreateContent($"RSS Item {m.Id}", 1061, "RssItem");
+                    rssItem.SetValue("title", m.Title);
+                    rssItem.SetValue("description", m.Description);
+                    rssItem.SetValue("categories", m.Categories);
+                    rssItem.SetValue("date", m.Date);
+                    rssItem.SetValue("link", m.Link);
+                    return rssItem;
+                });
+
+                Services.ContentService.DeleteContentOfType(1060);
+
+                foreach (IContent rssItem in rssItems)
+                {
+                    Services.ContentService.SaveAndPublishWithStatus(rssItem);
+                }
+
+                TempData.Add("Success", $"URL {rssFeedUrl.Url} was successfully imported.");
+                return RedirectToCurrentUmbracoPage();
+            }
+            catch (RssReaderException e)
+            {
+                TempData.Add("Failure", $"{e.Message} URL: {e.Url}");
+                return CurrentUmbracoPage();
+            }
         }
     }
 }
